+++ TechTalk.SpecFlow.Rpc/Shared/Request/Request.cs
@@ -7,5 +7,6 @@
         public string Method { get; set; }
         public string Arguments { get; set; }
         public bool IsShutDown { get; set; }
+        public bool IsPing { get; set; }
     }
 }
