+++ TechTalk.SpecFlow.Utils/TechTalk.SpecFlow.Utils.csproj
@@ -14,5 +14,7 @@
     <GenerateAssemblyFileVersionAttribute>false</GenerateAssemblyFileVersionAttribute>
   </PropertyGroup>
 
+  
+
 
 </Project>
